moves collection screens to new a sub dir of collections
