+++ src/lib/Scenes/Collection/Screens/__tests__/CollectionHeader-tests.tsx
@@ -2,7 +2,7 @@ import { mount } from "enzyme"
 import OpaqueImageView from "lib/Components/OpaqueImageView/OpaqueImageView"
 import React from "react"
 import * as renderer from "react-test-renderer"
-import { CollectionFixture } from "../__fixtures__/index"
+import { CollectionFixture } from "../../Components/__fixtures__/index"
 import { CollectionHeader } from "../CollectionHeader"
 
 describe("CollectionHeader", () => {
